-  gestion de base des mots des listes (ajout et suppression) pour le ROLE_ADMINISTRATEUR

+++ corpus/src/main/java/com/servicelibre/entities/corpus/ListeMot.java
@@ -25,7 +25,7 @@ import javax.persistence.UniqueConstraint;
 public class ListeMot {
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	int Id;
+	long Id;
 
 	@ManyToOne(optional = false)
 	Mot mot;
@@ -50,11 +50,11 @@ public class ListeMot {
 		this.note = "";
 	}
 
-	public int getId() {
+	public long getId() {
 		return Id;
 	}
 
-	public void setId(int id) {
+	public void setId(long id) {
 		Id = id;
 	}
